oil_prices/requirements.txt
amqp==5.0.6
asgiref==3.3.4
backcall==0.2.0
billiard==3.6.4.0
celery==5.1.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cycler==0.10.0
decorator==4.4.2
Django==3.2.3
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.4
fredapi==0.4.3
idna==2.10
inflection==0.5.1
ipykernel==5.5.0
ipython==7.21.0
ipython-genutils==0.2.0
jedi==0.18.0
joblib==1.0.1
jupyter-client==6.1.11
jupyter-core==4.7.1
kiwisolver==1.3.1
kombu==5.1.0
lxml==4.6.2
Markdown==3.3.4
matplotlib==3.3.4
more-itertools==8.7.0
numpy==1.20.1
pandas==1.2.3
pandas-datareader==0.9.0
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
prompt-toolkit==3.0.16
# psycopg2==2.8.6
Pygments==2.8.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
pywin32==300
pyzmq==22.0.3
Quandl==3.6.1
redis==3.5.3
requests==2.25.1
scikit-learn==0.24.1
scipy==1.6.2
six==1.15.0
sklearn==0.0
sqlparse==0.4.1
threadpoolctl==2.1.0
tornado==6.1
traitlets==5.0.5
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
